requirements.txt
numpy==1.26.0
albumentations==0.4.3
datasets==3.1.0
diffusers==0.35.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
pudb==2019.2
invisible-watermark
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
torch-fidelity==0.3.0
transformers==4.26.1
torchmetrics==0.6.0
kornia==0.6.0
openai
gradio
seaborn
Pillow
scipy
matplotlib
tqdm
fairscale==0.4.13
ftfy==6.3.1
accelerate==1.10.0
clip-anytorch==2.6.0
git+https://github.com/CompVis/taming-transformers.git@master
git+https://github.com/openai/CLIP.git@main
git+https://github.com/crowsonkb/k-diffusion.git